+++ squashfsspec/__init__.py
@@ -1,4 +1,7 @@
 # First-party
-from squashfsspec.squashfsspec import OffsetWrapper, SquashFSFileSystem
+from squashfsspec.squashfsspec import (
+    OffsetWrapper,
+    SquashFSFileSystem,
+)
 
 __all__ = ["SquashFSFileSystem", "OffsetWrapper"]